Split into patches
