+++ src/Addons/34thPRC_ArmourStandard/data/config_vehicles.hpp
@@ -1,5 +1,4 @@
 #include "marines2\config_vehicles.hpp"
 #include "mkiv\config_vehicles.hpp"
 #include "odst\config_vehicles.hpp"
-#include "odst2\config_vehicles.hpp"
 #include "mkvb\config_vehicles.hpp"
