--- src/Addons/34thPRC_ArmourStandard/data/odst2/backpacks/config_vehicles.hpp
@@ -1,1 +1,0 @@
-#include "s9solajumpjet\config_vehicles.hpp"
