--- src/Addons/34thPRC_ArmourStandard/data/odst2/config_vehicles.hpp
@@ -1,1 +1,0 @@
-#include "backpacks\config_vehicles.hpp"
